Moved windowing menu initialization to action handler.
Additionally, fixed circular import crash.
