+++ src/View/mainwindowrestructure/NewMenuBar.py
@@ -33,10 +33,6 @@ class NewMenuBar(QtWidgets.QMenuBar):
 
         self.addAction(self.action_help)
 
-        # Create sub-menu for Windowing
-        self.menu_windowing = QtWidgets.QMenu(self.menu_tools)
-        self.init_windowing_menu()
-
         # Add actions to File menu
         self.menu_file.addAction(self.action_handler.action_open)
         self.menu_file.addSeparator()
@@ -47,40 +43,10 @@ class NewMenuBar(QtWidgets.QMenuBar):
         # Add actions to Tool menu
         self.menu_tools.addAction(self.action_handler.action_zoom_in)
         self.menu_tools.addAction(self.action_handler.action_zoom_out)
-        self.menu_tools.addMenu(self.menu_windowing)
+        self.menu_tools.addMenu(self.action_handler.menu_windowing)
         self.menu_tools.addAction(self.action_handler.action_transect)
         self.menu_tools.addAction(self.action_handler.action_add_ons)
 
         # Add actions to Export menu
         self.menu_export.addAction(self.action_handler.action_clinical_data_export)
         self.menu_export.addAction(self.action_handler.action_pyradiomics_export)
-
-    def init_windowing_menu(self):
-        icon_windowing = QtGui.QIcon()
-        icon_windowing.addPixmap(
-            QtGui.QPixmap(":/images/Icon/windowing.png"),
-            QtGui.QIcon.Normal,
-            QtGui.QIcon.On
-        )
-        self.menu_windowing.setIcon(icon_windowing)
-        self.menu_windowing.setTitle("Windowing")
-
-        patient_dict_container = PatientDictContainer()
-        dict_windowing = patient_dict_container.get("dict_windowing")
-
-        # Get the right order for windowing names
-        names_ordered = sorted(dict_windowing.keys())
-        if "Normal" in dict_windowing.keys():
-            old_index = names_ordered.index("Normal")
-            names_ordered.insert(0, names_ordered.pop(old_index))
-
-        # Create actions for each windowing item
-        for name in names_ordered:
-            text = str(name)
-            print(text)
-            action_windowing_item = QtWidgets.QAction()
-            action_windowing_item.triggered.connect(
-                lambda state, text=name: self.action_handler.windowing_handler(state, text)
-            )
-            action_windowing_item.setText(text)
-            self.menu_windowing.addAction(action_windowing_item)
